refactor: Move CsvFileGetter to CsvFileService and move security check into service
