+++ HerPortal.UnitTests/BusinessLogic/Services/RegularJobsServiceTests.cs
@@ -1,10 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using HerPortal.BusinessLogic.ExternalServices.CsvFiles;
 using HerPortal.BusinessLogic.ExternalServices.EmailSending;
 using HerPortal.BusinessLogic.Models;
 using HerPortal.BusinessLogic.Services;
+using HerPortal.BusinessLogic.Services.CsvFileService;
 using HerPortal.Data;
 using Microsoft.Extensions.Logging;
 using NUnit.Framework;
@@ -19,7 +19,7 @@ public class RegularJobsServiceTests
     private Mock<ILogger<RegularJobsService>> mockLogger;
     private Mock<IDataAccessProvider> mockDataAccessProvider;
     private Mock<IEmailSender> mockEmailSender;
-    private Mock<ICsvFileGetter> mockCsvFileGetter;
+    private Mock<ICsvFileService> mockCsvFileService;
     private RegularJobsService underTest;
     
     private const string EmailAddress = "test@example.com";
@@ -30,9 +30,9 @@ public class RegularJobsServiceTests
         mockLogger = new Mock<ILogger<RegularJobsService>>();
         mockDataAccessProvider = new Mock<IDataAccessProvider>();
         mockEmailSender = new Mock<IEmailSender>();
-        mockCsvFileGetter = new Mock<ICsvFileGetter>();
+        mockCsvFileService = new Mock<ICsvFileService>();
         
-        underTest = new RegularJobsService(mockDataAccessProvider.Object, mockEmailSender.Object, mockCsvFileGetter.Object, mockLogger.Object);
+        underTest = new RegularJobsService(mockDataAccessProvider.Object, mockEmailSender.Object, mockCsvFileService.Object, mockLogger.Object);
     }
     
     [Test]
@@ -75,7 +75,7 @@ public class RegularJobsServiceTests
                 .Build()
         };
         mockDataAccessProvider.Setup(dap => dap.GetAllActiveUsersAsync()).ReturnsAsync(users);
-        mockCsvFileGetter.Setup(cfg => cfg.GetByCustodianCodesAsync(new[] { "114" }, users[0].Id )).ReturnsAsync(files);
+        mockCsvFileService.Setup(cfg => cfg.GetByCustodianCodesAsync(new[] { "114" }, users[0].Id )).ReturnsAsync(files);
         
         // Act
         await underTest.SendReminderEmailsAsync();
@@ -123,8 +123,8 @@ public class RegularJobsServiceTests
                 .Build()
         };
         mockDataAccessProvider.Setup(dap => dap.GetAllActiveUsersAsync()).ReturnsAsync(users);
-        mockCsvFileGetter.Setup(cfg => cfg.GetByCustodianCodesAsync(new[] { "114" }, users[0].Id )).ReturnsAsync(user1Files);
-        mockCsvFileGetter.Setup(cfg => cfg.GetByCustodianCodesAsync(new[] { "910" }, users[1].Id )).ReturnsAsync(user2Files);
+        mockCsvFileService.Setup(cfg => cfg.GetByCustodianCodesAsync(new[] { "114" }, users[0].Id )).ReturnsAsync(user1Files);
+        mockCsvFileService.Setup(cfg => cfg.GetByCustodianCodesAsync(new[] { "910" }, users[1].Id )).ReturnsAsync(user2Files);
         
         // Act
         await underTest.SendReminderEmailsAsync();
@@ -173,8 +173,8 @@ public class RegularJobsServiceTests
                 .Build()
         };
         mockDataAccessProvider.Setup(dap => dap.GetAllActiveUsersAsync()).ReturnsAsync(users);
-        mockCsvFileGetter.Setup(cfg => cfg.GetByCustodianCodesAsync(new[] { "114" }, users[0].Id )).ReturnsAsync(user1Files);
-        mockCsvFileGetter.Setup(cfg => cfg.GetByCustodianCodesAsync(new[] { "910" }, users[1].Id )).ReturnsAsync(user2Files);
+        mockCsvFileService.Setup(cfg => cfg.GetByCustodianCodesAsync(new[] { "114" }, users[0].Id )).ReturnsAsync(user1Files);
+        mockCsvFileService.Setup(cfg => cfg.GetByCustodianCodesAsync(new[] { "910" }, users[1].Id )).ReturnsAsync(user2Files);
         mockEmailSender.Setup(es => es.SendNewReferralReminderEmail("user1@example.com")).Throws(new EmailSenderException(EmailSenderExceptionType.Other));
         
         // Act
