+++ HerPortal.UnitTests/Website/Models/HomepageViewModelTests.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using FluentAssertions;
 using HerPortal.BusinessLogic.Models;
+using HerPortal.BusinessLogic.Services.CsvFileService;
 using HerPortal.Models;
 using NUnit.Framework;
-using CsvFileData = HerPortal.BusinessLogic.ExternalServices.CsvFiles.CsvFileData;
 using Tests.Helpers;
 
 namespace Tests.Website.Models;
